[M4] refactor: 클래스 네이밍 변경 (Records -> Record)
참고: https://edu.nextstep.camp/s/DVVUSLFe/ls/wNqIoaAF
