+++ src/main/java/bowling/domain/BowlingGame.java
@@ -2,35 +2,35 @@ package bowling.domain;
 
 public class BowlingGame {
     private final Player player;
-    private final Records records;
+    private final Record record;
 
     public BowlingGame(Player player) {
         this.player = player;
-        this.records = new Records();
+        this.record = new Record();
     }
 
     public boolean isEndGame() {
-        return records.isLastFrame() && records.isEndLastFrame();
+        return record.isLastFrame() && record.isEndLastFrame();
     }
 
     public void doGame(int downPinCount) {
-        records.record(downPinCount);
+        record.record(downPinCount);
     }
 
     public int getFrameRound() {
-        int recordCount = records.getRecordCount();
+        int recordCount = record.getRecordCount();
         if (recordCount == 0) {
             return 1;
         }
 
-        if (records.isEndLastFrame()) {
+        if (record.isEndLastFrame()) {
             return recordCount + 1;
         }
         return recordCount;
     }
 
-    public Records getRecords() {
-        return records;
+    public Record getRecord() {
+        return record;
     }
 
     public Player getPlayer() {
